src/main/java/no/cantara/jau/eventextraction/EventExtractorService.java
package no.cantara.jau.eventextraction;

import no.cantara.jau.serviceconfig.client.ConfigServiceClient;
import no.cantara.jau.serviceconfig.dto.EventExtractionConfig;
import no.cantara.jau.serviceconfig.dto.EventExtractionTag;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import java.util.List;
import java.util.Map;

import static java.util.stream.Collectors.*;

public class EventExtractorService {
    private static final Logger log = LoggerFactory.getLogger(EventExtractorService.class);
    private final EventRepo repo;
    private final ConfigServiceClient client;

    public EventExtractorService(EventRepo repo, ConfigServiceClient client) {
        this.repo = repo;
        this.client = client;
    }

    /**
     * Can probably be moved to configservice-sdk
     */
    public Map<String, List<EventExtractionTag>> groupExtractionConfigsByFile(
            EventExtractionConfig config) {
        Map<String, List<EventExtractionTag>> collect = config.tags.stream()
                .collect(groupingBy(item -> item.filePath));
        return collect;
    }
}
